+++ src/main/groovy/com/cedarsoftware/ncube/UrlCommandCell.groovy
@@ -67,15 +67,39 @@ abstract class UrlCommandCell implements CommandCell
         return cacheable
     }
 
+    // TODO: Restore this when L3 cache fully implemented
+//    void clearClassLoaderCache()
+//    {
+//        // classpath case, lets clear all classes before setting to null.
+//        def localVar = cache
+//        if (localVar instanceof GroovyClassLoader)
+//        {
+//            (localVar as GroovyClassLoader).clearCache()
+//        }
+//        cache = null
+//    }
+
     void clearClassLoaderCache()
     {
-        // classpath case, lets clear all classes before setting to null.
-        def localVar = cache
-        if (localVar instanceof GroovyClassLoader)
+        if (cache == null)
         {
-            (localVar as GroovyClassLoader).clearCache()
+            return
         }
-        cache = null
+
+        synchronized (GroovyBase.class)
+        {
+            if (cache == null)
+            {
+                return
+            }
+
+            // classpath case, lets clear all classes before setting to null.
+            if (cache instanceof GroovyClassLoader)
+            {
+                ((GroovyClassLoader)cache).clearCache()
+            }
+            cache = null
+        }
     }
 
     protected URL getActualUrl(Map<String, Object> ctx)