+++ src/test/groovy/com/cedarsoftware/ncube/TestNCube.groovy
@@ -3532,7 +3532,7 @@ class TestNCube
         coord["protocol"] = "http"
         coord["content"] = "hello2"
         def x = ncube.getCell(coord)
-        assert x == "Hello, world."  // Not 2, because the same class name was used for same classpath.
+        assert x == "Hello, world 2."  // Not 2, because the same class name was used for same classpath.
 
         coord["protocol"] = "http"
         coord["content"] = "95"